JSdoc, test types and separate request argument type file for dnd.* APIs
